feat: add frequency to prompt description
